Add files via upload
The short story game "New Year's Walkthrough". At first, I just thought that I would be playing mini-games. But then, when I started making the third mini-game, namely riddles, I realized that it was too boring and tedious. Then I decided to do this: after completing some puzzles, a mini-game would open. It took some time, and it happened by accident, but slowly, turned into a novel. At the moment, the game is not very well done, I just did not have time to finish it before the required date (in short, our group was given a chance to make the game in mathematics, since we have a group of IT specialists). I've done what I've done. Maybe I'll redo it in a beautiful way later, but for now. Happy New Year 2025 to all of you!